axp-op-12-test: use strict types for mask calculation

+++ axp-op-12-test.c
@@ -114,17 +114,17 @@ static inline uint64_t axp_ins_pm (int f, uint64_t a, uint64_t b)
 	return axp_sr (f, axp_zap (a, ~axp_byte_mask (f)), b * 8, 0);
 }
 
-static uint64_t axp_ms (int f, uint64_t b, int p)
+static uint8_t axp_ms (int f, uint64_t b, int p)
 {
 	const int x = F0 &  p;		/* preinvert byte mask		*/
 	const int h = F6 & !F3;		/* select high mask		*/
 	const int y = F0 |  p;		/* shift byte mask		*/
 	const int z = F0 | !p;		/* postinvert byte mask		*/
 
-	unsigned m  = axp_byte_mask (f);
-	unsigned mn = x ? m ^ 0x00ff : m;  /* invert whole 8-bit mask */
-	unsigned sm = h ? mn << (b & 7) >> 8 : mn << (b & 7);
-	unsigned ms = y ? sm : m;
+	uint16_t m  = axp_byte_mask (f);
+	uint16_t mn = x ? m ^ 0x00ff : m;  /* invert whole 8-bit mask */
+	uint8_t  sm = h ? mn << (b & 7) >> 8 : mn << (b & 7);
+	uint8_t  ms = y ? sm : m;
 
 	return z ? ~ms : ms;
 }
@@ -134,7 +134,7 @@ static inline uint64_t axp_mei (int f, uint64_t a, uint64_t bs, uint64_t bm)
 	const int p = !(F2 | F3);	/* pass A (zap/msk block)	*/
 
 	const uint64_t as = axp_sr (f, a, F1 ? bs * 8 : bs, p);
-	const unsigned ms = axp_ms (f, bm, p);
+	const uint8_t  ms = axp_ms (f, bm, p);
 
 	return axp_zap (as, F1 ? ms : !p ? 0 : F0 ? ~bm : bm);
 }
